fix(analysis): address review findings
Real bugs:
- Stamp analyzer model on pending rows so the detail metadata bar isn't
  blank (worker.Enqueue defaults req.Model from runner.Model())
- Use context.WithoutCancel for the queue-full rollback delete so a
  cancelled HTTP request can't orphan a pending row
- Change SeverityLevelComparison.Current to float64; multi-day periods
  no longer truncate sub-1.0 daily averages to zero (prompt now emits
  "%.1f/day" for both current and baseline)
- Scheduler snaps period_end to the schedule's wall-clock time, so
  retries within the 5-minute firing window resolve to the same window
  and the duplicate-active index actually dedupes them
- usePolling guards against re-entry with a runId counter so a second
  start() while a previous tick is in flight no longer leaks a parallel
  timer chain; stop() also invalidates in-flight ticks
- Print stylesheet hides every direct sibling of <main> via
  body > div > *:not(main), matching the actual App.vue shell instead
  of guessing at nav/header/aside selectors
- Replace stale openapi.yml entries (latest/{id}/trigger) with the
  current routes and add AnalysisSchedule schema + schedule CRUD paths

Polish:
- AnalysisScheduler.RunNow uses GetAnalysisSchedule instead of listing
  all schedules and filtering
- Modal binds Escape to close while open (PRD requirement)
- runResults keyed by schedule id so triggering several rows doesn't
  silently overwrite earlier outcomes; explicit scheduler_disabled
  message
- Collapse duplicate scanAnalysisSchedule/Row into one helper taking
  pgx.Row (pgx.Rows satisfies the interface)
- Extract feedBadgeClass/statusBadgeClass/formatDate/timeAgo/
  formatDuration into lib/analysis-format.ts; both Vue views import
- AnalysisReports/Detail use polling.start() as the initial fetch
  instead of a separate refresh() call, avoiding the double request on
  mount
- UTF-8-safe truncateErr trims back to a rune boundary

+++ api/internal/model/analysis.go
@@ -90,11 +90,13 @@ type HostErrorCount struct {
 	TopMsgID string `json:"top_msgid"`
 }
 
-// SeverityLevelComparison compares current severity count to baseline average.
+// SeverityLevelComparison compares current severity rate to baseline average.
+// Both Current and BaselineAvg are per-day rates so multi-day periods compare
+// apples-to-apples; for a 24h run, Current equals the raw count.
 type SeverityLevelComparison struct {
 	Severity    int     `json:"severity"`
 	Label       string  `json:"label"`
-	Current     int64   `json:"current"`
+	Current     float64 `json:"current"`
 	BaselineAvg float64 `json:"baseline_avg"`
 	ChangePct   float64 `json:"change_pct"`
 }
